Give libpas a makefile (with autodeps and everything) rather than using a shell script. Also, fix two bugs in local allocator in the case where the bmalloc call to stash alloc bits causes a TLC resize.

+++ filc/main/main.c
@@ -32,7 +32,7 @@ int main(int argc, char** argv)
     int* exit_args;
     filc_ptr pizlonated_argv;
     int index;
-    size_t environ_size;
+    int environ_size;
     filc_ptr environ_ptr;
     filc_ptr __init_libc_ptr;
     filc_ptr main_ptr;
